Cleanup Metadata usages
